Controller implementation is split into Interface and implementation.
