+++ io.openems.edge.controller.ess.limittotaldischarge/src/io/openems/edge/controller/ess/limittotaldischarge/Config.java
@@ -6,7 +6,8 @@ import org.osgi.service.metatype.annotations.ObjectClassDefinition;
 @ObjectClassDefinition(//
 		name = "Controller Ess Limit Total Discharge", //
 		description = "Limits total discharge for an Ess.")
-@interface Config {
+
+public @interface Config {
 
 	@AttributeDefinition(name = "Component-ID", description = "Unique ID of this Component")
 	String id() default "ctrlLimitTotalDischarge0";
